🚚 move: 패키지 구조 정리 및 의존성 개선 (#57)
- 서비스 패키지 구조를 명확히 하기 위해 facade 패키지 내의 클래스를 command 및 query로 분리
