+++ src/main/java/troublog/backend/domain/trouble/service/facade/command/CommentCommandFacade.java
@@ -1,4 +1,4 @@
-package troublog.backend.domain.trouble.service.facade;
+package troublog.backend.domain.trouble.service.facade.command;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,6 +11,7 @@ import troublog.backend.domain.trouble.dto.response.CommentResDto;
 import troublog.backend.domain.trouble.entity.Comment;
 import troublog.backend.domain.trouble.entity.Post;
 import troublog.backend.domain.trouble.service.command.CommentCommandService;
+import troublog.backend.domain.trouble.service.facade.relation.CommentRelationFacade;
 import troublog.backend.domain.trouble.service.factory.CommentFactory;
 import troublog.backend.domain.trouble.service.factory.PostFactory;
 import troublog.backend.domain.trouble.service.query.CommentQueryService;
