+++ src/main/java/troublog/backend/domain/trouble/service/facade/command/LikeCommandFacade.java
@@ -1,4 +1,4 @@
-package troublog.backend.domain.trouble.service.facade;
+package troublog.backend.domain.trouble.service.facade.command;
 
 import java.util.Collections;
 import java.util.List;
