+++ src/main/java/troublog/backend/domain/trouble/service/facade/query/CommentQueryFacade.java
@@ -1,4 +1,4 @@
-package troublog.backend.domain.trouble.service.facade;
+package troublog.backend.domain.trouble.service.facade.query;
 
 import java.util.Collections;
 import java.util.List;
