+++ src/main/java/troublog/backend/domain/trouble/service/facade/query/PostQueryFacade.java
@@ -1,4 +1,4 @@
-package troublog.backend.domain.trouble.service.facade;
+package troublog.backend.domain.trouble.service.facade.query;
 
 import java.util.List;
 import java.util.Map;
