+++ src/main/java/troublog/backend/domain/trouble/service/facade/relation/PostRelationFacade.java
@@ -1,4 +1,4 @@
-package troublog.backend.domain.trouble.service.facade;
+package troublog.backend.domain.trouble.service.facade.relation;
 
 import java.util.List;
